+++ source/modules/game/audio/audioEvent.h
@@ -31,7 +31,7 @@ public:
 
     bool getPaused() const;
     float getVolume() const;
-    float getLength() const;
+    int getLength() const;
     float getPitch() const;
     float getParameter(const std::string& name);
